requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1
torchvision==0.17.2
torchaudio==2.2.2
fastapi==0.128.2
uvicorn==0.40.0
qdrant-client==1.16.2
sentence-transformers==2.7.0
openai==2.17.0
python-dotenv==1.2.1
pydantic==2.12.5
lxml==6.0.2
httpx==0.28.1
rank-bm25
rapidfuzz
requests
numpy
